# lib/api-zod/src/generated/types/channel.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Channel {
  id: number;
  slug: string;
  name: string;
  description: string;
  createdAt: Date;
}