# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MessageStatus } from './messageStatus';
import type { Reactions } from './reactions';

export interface Message {
  id: number;
  channelId: number;
  username: string;
  content: string;
  status: MessageStatus;
  reactions: Reactions;
  createdAt: Date;
}