# lib/api-zod/src/generated/types/reactionInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReactionInput {
  /**
     * @minLength 1
     * @maxLength 32
     */
  username: string;
  /**
     * @minLength 1
     * @maxLength 8
     */
  emoji: string;
}