# lib/api-zod/src/generated/types/reactions.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Map of emoji to the usernames who reacted with it
 */
export interface Reactions {[key: string]: string[]}