# lib/api-zod/src/generated/types/signupBody.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SignupBody {
  /**
     * @minLength 2
     * @maxLength 32
     */
  username: string;
  /**
     * @minLength 4
     * @maxLength 128
     */
  password: string;
  /**
     * @minLength 1
     * @maxLength 64
     */
  displayName: string;
}