# lib/db/src/schema/messages.ts
import {
  integer,
  jsonb,
  pgTable,
  serial,
  text,
  timestamp,
  varchar,
} from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";
import { channelsTable } from "./channels";

/** emoji -> usernames who reacted with it */
export type Reactions = Record<string, string[]>;

export const messagesTable = pgTable("messages", {
  id: serial("id").primaryKey(),
  channelId: integer("channel_id")
    .notNull()
    .references(() => channelsTable.id),
  username: varchar("username", { length: 32 }).notNull(),
  content: text("content").notNull(),
  status: varchar("status", { length: 16 }).notNull().default("sent"),
  reactions: jsonb("reactions").$type<Reactions>().notNull().default({}),
  createdAt: timestamp("created_at", { withTimezone: true })
    .notNull()
    .defaultNow(),
});

export const insertMessageSchema = createInsertSchema(messagesTable).omit({
  id: true,
  createdAt: true,
  status: true,
  reactions: true,
});
export type InsertMessage = z.infer<typeof insertMessageSchema>;
export type Message = typeof messagesTable.$inferSelect;